// client/src/context/actions.tsx
export enum Action {
    DISPLAY_ALERT = 'SHOW_ALERT',
    CLEAR_ALERT = 'CLEAR_ALERT',
    SETUP_USER_BEGIN = 'SETUP_USER_BEGIN',
    SETUP_USER_SUCCESS = 'SETUP_USER_SUCCESS',
    SETUP_USER_ERROR = 'SETUP_USER_ERROR',
    TOGGLE_SIDEBAR = 'TOGGLE_SIDEBAR',
    LOGOUT_USER = 'LOGOUT_USER',
    UPDATE_USER_BEGIN = 'UPDATE_USER_BEGIN',
    UPDATE_USER_SUCCESS = 'UPDATE_USER_SUCCESS',
    UPDATE_USER_ERROR = 'UPDATE_USER_ERROR',
    HANDLE_CHANGE = 'HANDLE_CHANGE',
    CLEAR_VALUES = 'CLEAR_VALUES',
    CREATE_JOB_BEGIN = 'CREATE_JOB_BEGIN',
    CREATE_JOB_SUCCESS = 'CREATE_JOB_SUCCESS',
    CREATE_JOB_ERROR = 'CREATE_JOB_ERROR',
    GET_JOBS_BEGIN = 'GET_JOBS_BEGIN',
    GET_JOBS_SUCCESS = 'GET_JOBS_SUCCESS',
    SET_EDIT_JOB = 'SET_EDIT_JOB',
    LEAVE_EDIT_PAGE = 'LEAVE_EDIT_PAGE',
    DELETE_JOB_BEGIN = 'DELETE_JOB_BEGIN',
    EDIT_JOB_BEGIN = 'EDIT_JOB_BEGIN',
    EDIT_JOB_SUCCESS = 'EDIT_JOB_SUCCESS',
    EDIT_JOB_ERROR = 'EDIT_JOB_ERROR',
    SHOW_STATS_BEGIN = 'SHOW_STATS_BEGIN',
    SHOW_STATS_SUCCESS = 'SHOW_STATS_SUCCESS',
    CLEAR_FILTERS = 'CLEAR_FILTERS',
    CHANGE_PAGE = 'CHANGE_PAGE',
}

export enum JobTypes {
    FULL_TIME = 'Full-time',
    PART_TIME = 'Part-time',
    REMOTE = 'Remote',
    INTERNSHIP = 'Internship'
}

export enum StatusOptions {
    PENDING = 'Pending',
    INTERVIEW = 'Interview',
    DECLINED = 'Declined'
}

export enum sortOptions {
    LATEST = 'Latest',
    OLDEST = 'Oldest',
    A_Z = 'A-Z',
    Z_A = 'Z-A'
}